refactor(urlsubmission): nest remaining root types into folders by sub-namespace
Move the 22 remaining root-level UrlSubmission types into technical
folders matching house style, mirroring Persistence.Abstractions (#902)
and the existing Categorisation/Extensions/Factories/Models folders:

- Processors/: CategorisedItemProcessor, PodcastProcessor, DiscoveryResultProcessor (+ interfaces)
- Enrichers/: EpisodeEnricher, DescriptionHelper (+ interfaces)
- Matching/: EpisodeHelper (+ interface)
- Adaptors/: SubmitResultAdaptor (+ interface)
- Services/: PodcastService (+ interface)
- Submitters/: UrlSubmitter, DiscoveryUrlSubmitter (+ interfaces)
- Models/: SubmitResultState, DiscoverySubmitResultState (enums, alongside existing DTOs)

Constants.cs stays at root as a small cross-cutting leftover, matching the
same pattern already used in PodcastServices.Abstractions and Episodes.

Root .cs count: 23 -> 1 (Constants.cs). Total files unchanged at 45.

Public type names and DI extension method names (AddUrlSubmission) are
unchanged. Updated all consumer usings across Cloud/Api, Console-Apps,
and RedditPodcastPoster.UrlSubmission.Tests to import the new
sub-namespaces explicitly, since several previously relied on C#'s
enclosing-namespace fallback lookup that no longer applies once types
move out of the assembly root namespace.
