+++ Class-Libraries/RedditPodcastPoster.UrlSubmission.Tests/EpisodeHelperTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using Moq.AutoMock;
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.UrlSubmission.Categorisation;
+using RedditPodcastPoster.UrlSubmission.Matching;
 using RedditPodcastPoster.UrlSubmission.Models;
 using Xunit;
