+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Enrichers/EpisodeEnricher.cs
@@ -7,7 +7,7 @@ using RedditPodcastPoster.Models;
 using RedditPodcastPoster.UrlSubmission.Categorisation;
 using RedditPodcastPoster.UrlSubmission.Models;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Enrichers;
 
 public class EpisodeEnricher(
     IDescriptionHelper descriptionHelper,
