+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Enrichers/IDescriptionHelper.cs
@@ -1,6 +1,6 @@
 ﻿using RedditPodcastPoster.UrlSubmission.Categorisation;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Enrichers;
 
 public interface IDescriptionHelper
 {
