+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Enrichers/IEpisodeEnricher.cs
@@ -2,7 +2,7 @@
 using RedditPodcastPoster.UrlSubmission.Categorisation;
 using RedditPodcastPoster.UrlSubmission.Models;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Enrichers;
 
 public interface IEpisodeEnricher
 {
