+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Processors/DiscoveryResultProcessor.cs
@@ -12,7 +12,7 @@ using RedditPodcastPoster.UrlSubmission.Models;
 using RedditPodcastPoster.PodcastServices.Abstractions.Models;
 using Podcast = RedditPodcastPoster.Models.Podcast;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Processors;
 
 public class DiscoveryResultProcessor(
     IUrlCategoriser urlCategoriser,
