+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Processors/IPodcastProcessor.cs
@@ -1,7 +1,7 @@
 ﻿using RedditPodcastPoster.UrlSubmission.Categorisation;
 using RedditPodcastPoster.UrlSubmission.Models;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Processors;
 
 public interface IPodcastProcessor
 {
