+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Services/IPodcastService.cs
@@ -2,7 +2,7 @@
 using RedditPodcastPoster.PodcastServices.Abstractions;
 using RedditPodcastPoster.PodcastServices.Abstractions.Models;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Services;
 
 public interface IPodcastService
 {
