+++ Class-Libraries/RedditPodcastPoster.UrlSubmission/Submitters/IDiscoveryUrlSubmitter.cs
@@ -3,7 +3,7 @@ using RedditPodcastPoster.PodcastServices.Abstractions;
 using RedditPodcastPoster.UrlSubmission.Models;
 using RedditPodcastPoster.PodcastServices.Abstractions.Models;
 
-namespace RedditPodcastPoster.UrlSubmission;
+namespace RedditPodcastPoster.UrlSubmission.Submitters;
 
 public interface IDiscoveryUrlSubmitter
 {
