+++ Cloud/Api/Handlers/DiscoveryCurationHandler.cs
@@ -8,8 +8,8 @@ using RedditPodcastPoster.Auth0;
 using RedditPodcastPoster.EntitySearchIndexer;
 using RedditPodcastPoster.Episodes;
 using RedditPodcastPoster.PodcastServices.Abstractions;
-using RedditPodcastPoster.UrlSubmission;
 using RedditPodcastPoster.UrlSubmission.Models;
+using RedditPodcastPoster.UrlSubmission.Submitters;
 using RedditPodcastPoster.PodcastServices.Abstractions.Models;
 
 namespace Api.Handlers;
